Attempt to fix compile
