+++ federation/src/main/scala/caliban/federation/v2x/FederationV2.scala
@@ -1,12 +1,46 @@
 package caliban.federation.v2x
 
+import caliban.GraphQL
 import caliban.federation.{ FederationDirectives, FederationSupport }
+import caliban.rendering.{ DocumentRenderer, Renderer }
 
-class FederationV2(private[v2x] val extensions: List[Extension])
+class FederationV2(extensions: List[Extension])
     extends FederationSupport(Nil, extensions.map(_.toDirective))
     with FederationDirectives
-    with FederationDirectivesV2
-    with FederationRenderingV2
+    with FederationDirectivesV2 {
+
+  /**
+   * Constructs a renderer that can render a GraphQL schema with federation directives.
+   *
+   * This is useful if you need to render the schema to a string or file in CI or for debugging purposes.
+   *
+   * @note Make sure to use this renderer on the graph _before_ apply the federation aspect, otherwise it will include
+   *       the federation specific fields as well.
+   *
+   * @see [[renderFederatedCompact]] when you want to render a compact version of the schema.
+   */
+  def renderFederated[R](graphql: GraphQL[R]): String =
+    renderer[R].render(graphql)
+
+  /**
+   * Constructs a renderer that can render a GraphQL schema with federation directives in a compact format.
+   *
+   * This is useful if you need to render the schema to a string or file in CI or for debugging purposes.
+   *
+   * @note Make sure to use this renderer on the graph _before_ apply the federation aspect, otherwise it will include
+   *       the federation specific fields as well.
+   *
+   * @see [[renderFederated]] when you want to render the schema in pretty form.
+   */
+  def renderFederatedCompact[R](graphql: GraphQL[R]): String =
+    renderer[R].renderCompact(graphql)
+
+  private def renderer[R]: Renderer[GraphQL[R]] =
+    DocumentRenderer
+      .contramap[GraphQL[_]] {
+        _.withSchemaDirectives(extensions.map(_.toDirective)).toDocument
+      }
+}
 
 object FederationV2 {
