+++ federation/src/test/scala/caliban/federation/v2x/FederationV2Spec.scala
@@ -18,20 +18,6 @@ import zio.test._
 import java.util.UUID
 
 object FederationV2Spec extends ZIOSpecDefault {
-  object SchemaWithEntity {
-    import caliban.federation.v2_3._
-
-    @GQLKey("id")
-    @GQLShareable
-    case class Entity(
-      id: UUID
-    )
-
-    case class Query(
-      hello: String,
-      entity: Entity
-    )
-  }
 
   override def spec =
     suite("FederationV2Spec")(
@@ -252,18 +238,28 @@ object FederationV2Spec extends ZIOSpecDefault {
       },
       test("renderer renders the schema including the extensions") {
         import caliban.federation.v2_3._
-        import SchemaWithEntity.{ Query, Entity }
+
+        @GQLKey("id")
+        @GQLShareable
+        case class User(
+          id: UUID
+        )
+
+        case class Query(
+          hello: String,
+          user: User
+        )
 
         val api = graphQL(
           RootResolver(
-            Query(hello = "Hello World!", entity = Entity(UUID.randomUUID()))
+            Query(hello = "Hello World!", user = User(UUID.randomUUID()))
           )
         )
 
         assertTrue(
           renderFederatedCompact(
             api
-          ) == """schema@link(url:"https://specs.apollo.dev/federation/v2.3",import:["@key","@requires","@provides","@external","@shareable","@tag","@inaccessible","@override","@extends","@composeDirective","@interfaceObject"]){query:Query}type Entity@key(fields:"id")@shareable{id:ID!}type Query{hello:String! entity:Entity!}"""
+          ) == """schema@link(url:"https://specs.apollo.dev/federation/v2.3",import:["@key","@requires","@provides","@external","@shareable","@tag","@inaccessible","@override","@extends","@composeDirective","@interfaceObject"]){query:Query}type Query{hello:String! user:User!}type User@key(fields:"id")@shareable{id:ID!}"""
         )
 
       }
